DROOLS-6972 Refactor immutable compilation phases
